+++ server/routers/conference/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
 import { RouterContext } from 'react-router';
-import routes from '../../shared/routes/conference-routes';
-import { COMMUNITY_URL } from '../constants';
-import { useRoutes } from '../routes';
+import routes from '../../../shared/routes/conference-routes';
+import { COMMUNITY_URL } from '../../constants';
+import { useRoutes } from '../../routes';
 
 function sendSite(res, renderProps) {
   const content = renderToString(<RouterContext {...renderProps} />);
